move write storage util into codegen
