+++ src/db.rs
@@ -45,7 +45,7 @@ impl Db {
     }
 
     pub fn write_storage(&mut self, address: Address, key: U256, value: U256) {
-        let account = self.accounts.entry(address).or_insert(U256::zero());
+        let account = self.accounts.entry(address).or_default();
         account.storage.insert(key, value);
     }
