+++ src/utils.rs
@@ -952,73 +952,6 @@ pub(crate) fn compute_memory_cost<'c>(
     Ok(memory_cost)
 }
 
-// /// Wrapper for calling the [`write_storage`](crate::syscall::SyscallContext::write_storage) syscall.
-pub(crate) fn write_storage<'c>(
-    op_ctx: &'c OperationCtx,
-    block: &'c Block,
-    key: Value<'c, 'c>,
-    value: Value<'c, 'c>,
-) -> Result<(), CodegenError> {
-    let context = op_ctx.mlir_context;
-    let location = Location::unknown(context);
-    let uint256 = IntegerType::new(context, 256);
-    let ptr_type = pointer(context, 0);
-    let pointer_size = block
-        .append_operation(arith::constant(
-            context,
-            IntegerAttribute::new(uint256.into(), MAX_STACK_SIZE as i64).into(),
-            location,
-        ))
-        .result(0)?
-        .into();
-
-    // get the address of the key parameter
-    let key_ptr = block
-        .append_operation(llvm::alloca(
-            context,
-            pointer_size,
-            ptr_type,
-            location,
-            AllocaOptions::new().elem_type(Some(TypeAttribute::new(uint256.into()))),
-        ))
-        .result(0)?
-        .into();
-
-    let res = block.append_operation(llvm::store(
-        context,
-        key,
-        key_ptr,
-        location,
-        LoadStoreOptions::default(),
-    ));
-    assert!(res.verify());
-
-    // get the address of the value parameter
-    let value_ptr = block
-        .append_operation(llvm::alloca(
-            context,
-            pointer_size,
-            ptr_type,
-            location,
-            AllocaOptions::new().elem_type(Some(TypeAttribute::new(uint256.into()))),
-        ))
-        .result(0)?
-        .into();
-
-    let res = block.append_operation(llvm::store(
-        context,
-        value,
-        value_ptr,
-        location,
-        LoadStoreOptions::default(),
-    ));
-    assert!(res.verify());
-
-    op_ctx.storage_write_syscall(block, key, value, location);
-
-    Ok(())
-}
-
 /// Wrapper for calling the [`extend_memory`](crate::syscall::SyscallContext::extend_memory) syscall.
 /// Extends memory only if the current memory size is less than the required size, consuming the corresponding gas.
 pub(crate) fn extend_memory<'c>(
